FastAPI-Backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
firebase-admin